+++ src/main/java/com/training/controllers/HomeController.java
@@ -2,14 +2,20 @@ package com.training.controllers;
 
 import java.util.Optional;
 
+import javax.servlet.http.HttpSession;
+import javax.validation.Valid;
+
 import com.training.data.entity.Product;
 import com.training.data.repos.CategoryRepo;
 import com.training.data.repos.ProductRepo;
+import com.training.dto.MessageData;
 import com.training.dto.ProductData;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+ import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,6 +31,8 @@ public class HomeController {
     @Autowired
     private CategoryRepo categoryRepo;
 
+    @Autowired
+    private HttpSession session;
 
     @GetMapping
     public String main(Model model) {
@@ -43,14 +51,32 @@ public class HomeController {
     }
 
     @GetMapping("/product/add")
-    public String add(Model model) {
-        model.addAttribute("product", new ProductData());
-        model.addAttribute("categories", categoryRepo.findAll());
-        return "input";
+     public String add(Model model) {
+         if(session.getAttribute("CURRENT_USER")==null){
+             return "redirect:/users/login";
+         }
+         model.addAttribute("product", new ProductData());
+         model.addAttribute("categories", categoryRepo.findAll());
+         return "input";
     }
 
     @PostMapping("/product/save")
-    public String save(ProductData productData) {
+    public String save(@Valid ProductData productData, BindingResult resultErrors, Model model) {
+        
+        if(session.getAttribute("CURRENT_USER")==null){
+            return "redirect:/users/login";
+        }
+
+        if(resultErrors.hasErrors()) {
+            MessageData errorMessage = new MessageData();
+            for(ObjectError err: resultErrors.getAllErrors()) {
+                errorMessage.getMessages().add(err.getDefaultMessage());
+            }
+            model.addAttribute("product", productData);
+            model.addAttribute("ERROR", errorMessage);
+            model.addAttribute("categories", categoryRepo.findAll());
+            return "input";
+        }
         Product product = new Product();
         product.setCode(productData.getCode());
         product.setName(productData.getName());
@@ -63,12 +89,18 @@ public class HomeController {
 
     @GetMapping("/product/remove/{id}")
     public String remove(Model model, @PathVariable("id") Long id) {
+        if(session.getAttribute("CURRENT_USER")==null){
+            return "redirect:/users/login";
+        }
         repo.deleteById(id);
         return "redirect:/";
     }
 
     @GetMapping("/product/edit/{id}")
     public String edit(Model model, @PathVariable("id") Long id) {
+        if(session.getAttribute("CURRENT_USER")==null){
+            return "redirect:/users/login";
+        }
         Optional<Product> product = repo.findById(id);
         if(product.isPresent()) {
             model.addAttribute("product", product);
@@ -79,6 +111,9 @@ public class HomeController {
 
     @PostMapping("/product/update")
     public String update(Product product) {
+        if(session.getAttribute("CURRENT_USER")==null){
+            return "redirect:/users/login";
+        }
         repo.save(product);
         return "redirect:/";
     }
